+++ web/src/modules/Settings/components/_general/General/ShowEnginesSwitch/index.tsx
@@ -2,7 +2,11 @@ import { Switch } from "@mantine/core";
 import { useSettingsStore } from "@store/settings";
 import commonClasses from "../../../common/styles.module.scss";
 
-const ShowEnginesSwitch = () => {
+interface Props {
+  isM?: boolean;
+}
+
+const ShowEnginesSwitch: React.FC<Props> = ({ isM }) => {
   const showEngines = useSettingsStore((state) => state.showEngines);
   const setShowEngines = useSettingsStore((state) => state.setShowEngines);
 
@@ -11,6 +15,8 @@ const ShowEnginesSwitch = () => {
       className={commonClasses.settings_control}
       checked={showEngines}
       onChange={(e) => setShowEngines(e.currentTarget.checked)}
+      withThumbIndicator={isM ? false : true}
+      size={isM ? "md" : "sm"}
     />
   );
 };