+++ web/src/modules/SettingsMobile/components/common/utils.ts
@@ -1,0 +1,37 @@
+export const SEARCH_OPTIONS = [
+  // General
+  {
+    group: "General",
+    items: [
+      "Links in new tab",
+      "Website favicon",
+      "Search autocomplete",
+      "Show engines",
+      "Media in general",
+      "Instant Answers",
+      "Private Search",
+      // "Private Player",
+      // "Private View",
+    ],
+  },
+  {
+    group: "AI",
+    items: ["AI Answer", "AI Chat", "AI Summary"],
+  },
+
+  // Interface
+  {
+    group: "Categories",
+    items: ["Categories"],
+  },
+  {
+    group: "Interface",
+    items: ["Language", "Theme", "Primary color", "Color scheme"],
+  },
+
+  // Instances
+
+  // Engines
+
+  // Homepage
+];